# FlashD/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: jeefastis一个轻量级的Java快速开发平台
  version: 1.0.0
  title: jeefast快速开发平台
basePath: /jeefast
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义interfacedata
paths:
...
  /sys/dept/info/{deptId}:
    get:
      tags:
        - Department
      summary: get部门info
      produces:
        - application/json
      parameters:
        - name: deptId
          description: deptId
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: return部门info
          schema:
            $ref: '#/definitions/SysDept'
  /sys/dept/save:
    post:
      tags:
        - Department
      summary: add部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门object
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/update:
    post:
      tags:
        - Department
      summary: update部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门object
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/dept/delete:
    post:
      tags:
        - Department
      summary: delete部门
      produces:
        - application/json
      parameters:
        - name: body
          description: deptIdlist
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/list:
    get:
      tags:
        - Roles
      summary: rolelist
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: sort字段
          in: query
          type: string
        - name: order
          description: sort方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: role名
          in: query
          type: string
      responses:
        '200':
          description: returnmanage员list
          schema:
            $ref: '#/definitions/SysRole'
...
  /sys/role/info/{roleId}:
    get:
      tags:
        - Roles
      summary: getroleinfo
      produces:
        - application/json
      parameters:
        - name: roleId
          description: roleID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: returnroleinfo
          schema:
            $ref: '#/definitions/SysRole'
  /sys/role/save:
    post:
      tags:
        - Roles
      summary: addrole
      produces:
        - application/json
      parameters:
        - name: body
          description: roleobject
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/update:
    post:
      tags:
        - Roles
      summary: updaterole
      produces:
        - application/json
      parameters:
        - name: body
          description: roleobject
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysRoleEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/delete:
    post:
      tags:
        - Roles
      summary: deleterole
      produces:
        - application/json
      parameters:
        - name: body
          description: roleIDlist
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - menu management
      summary: getmenuinfo
      produces:
        - application/json
      parameters:
        - name: menuId
          description: menuID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: returnmenuinfo
          schema:
            $ref: '#/definitions/SysRole'
  /sys/menu/save:
    post:
      tags:
        - menu management
      summary: addmenu
      produces:
        - application/json
      parameters:
        - name: body
          description: menuobject
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
        - menu management
      summary: updatemenu
      produces:
        - application/json
      parameters:
        - name: body
          description: menuobject
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete:
    post:
      tags:
        - menu management
      summary: deletemenu
      produces:
        - application/json
      parameters:
        - name: body
          description: menuIDlist
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
      type: object
      properties:
        code:
          description: state码 0：success  非0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysDept:
    type: object
    properties:
      deptId:
        description: deptId
        type: integer
        format: int64
      name:
        description: Department Name
        type: string
      parentId:
        description: Superior departmentID，First level department为0
        type: integer
        format: int64
      parentName:
        description: superiorDepartment Name
        type: string
      orderNum:
        description: sort
        type: integer
        format: int32
      open:
        description: isno展开   true：展开   false：不展开
        type: boolean
        format: int32
  SysDeptEdit:
    type: object
    properties:
      name:
        description: Department Name
        type: string
      parentId:
        description: Superior departmentID，First level department为0
        type: integer
        format: int64
      orderNum:
        description: sort
        type: integer
        format: int32
  SysRole:
    type: object
    properties:
      roleId:
        description: roleID
        type: integer
        format: int64
      roleName:
        description: RoleName
        type: string
      deptId:
        description: deptId
        type: integer
        format: int64
      deptName:
        description: Department Name
        type: string
      remark:
        description: remark
        type: string
      menuIdList:
        description: menuIDlist
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: createUserID
        type: integer
        format: int64
      createTime:
        description: createTime
        type: string
        format: date-time
  SysRoleEdit:
    type: object
    properties:
      roleId:
        description: roleID
        type: integer
        format: int64
      roleName:
        description: RoleName
        type: string
      deptId:
        description: deptId
        type: integer
        format: int64
      remark:
        description: remark
        type: string
      menuIdList:
        description: menuIDlist
        type: array
        items:
          type: integer
          format: int64
...
  SysMenuEdit:
    type: object
    properties:
      menuId:
        description: menuID
        type: integer
        format: int64
      name:
        description: menu name
        type: string
      parentId:
        description: 父menuID，First level menu为0
        type: integer
        format: int64
      url:
        description: menuURL
        type: string
      perms:
        description: Authorization identification
        type: string
      type:
        description: type     0：directory   1：menu   2：button
        type: integer
        format: int32
      icon:
        description: menuicon
        type: string
      orderNum:
        description: sort
        type: integer
        format: int32
...